Split v2 metadata queries by resource
